\name{bog_subset}
\alias{bog_subset}

\title{
Subset OTU table by sampling site
}
\description{
Returns an OTU table containing only samples from the identified sampling site. This function can also be used on tables of higher level taxa generated by combine_otus().
}
\usage{
bog_subset(bog_id, table)
}
\arguments{
	\item{bog_id}{
  The three letter code indicating the sampling site. The bog is represented by letters one and two; options are TB, SS, CB, NS, MA, HK, WS, and FB.
The third letter indicates the layer; E for epilimnion and H for hypolimnion. Put quotes around the bog_id.
  }
	\item{table}{
  A table containing the relative abundances of each taxon as rows and samples as columns. Sample names must be coded in the format bog, layer, date, and replicate (example: TBE07JUN08.R2 == Trout Bog Epilimnion, collected 07Jun08, replicate 2)
  }
}


\value{
Returns a relative abundance table containing samples from a single sampling site in columns, with taxa in rows
}

\author{
Alexandra Linz <amlinz16@gmail.com>
}

\keyword{ data_processing }

\examples{
data(otu_table)

Trout_Bog_Epilimnion <- bog_subset("TBE", otu_table)

Hells_Kitchen_Hypolimnion <- bog_subset("HKH", otu_table)

}
