\name{metadata}
\alias{metadata}

\title{
Lake metadata for OTU table
}
\description{
A dataset containing temperature and oxygen profiles from the lakes in this study

}
\usage{
data(taxonomy)
}
\format{
A dataframe with 6 columns (measured variables) and 13,607 rows	(depth profiles)
}


\details{
Missing data indicated by NA
Some sample dates and metadata dates may not match up exactly; if this presents an issue, please email and I will look at our written records for the right date
Epilimnion and hypolimnion samples each have an identical depth profile entry; search for just one or the other
}

\author{
Alexandra Linz <amlinz16@gmail.com>
}

