\name{clean_taxonomy}
\alias{clean_taxonomy}

\title{
Clean Taxonomy File
}
\description{
Reduces information in a mothur .taxonomy file by removing the second column with the number of reads per OTU.
It will also check for and remove OTUs in the taxonomy file that are not in OTU table - this may be the case if rarefaction was performed after classification.
}
\usage{
clean_taxonomy(taxonomy_file, table)
}
\arguments{
	\item{taxonomy_file}{
  A .taxonomy file output by mothur
}
	\item{table}{
  An OTU table generated by clean_shared()
}
}

\value{
Returns the taxonomy with OTUs as row names and taxonomic assignments in column 1
}

\author{
Alexandra Linz <amlinz16@gmail.com>
}
\note{
Because of the order in the mothur workflow used in our paper, subsampling occurred after the taxonomy file was made. 

}

\keyword{ format_mothur_output }

\examples{

#Example path only: path <- "mothur_output/bogs.taxonomy"
#taxonomy <- clean_taxonomy(path)

}
