\name{OTUtable-package}
\alias{OTUtable-package}
\alias{OTUtable}
\docType{package}
\title{
OTU table analysis functions
}
\description{
This package contains functions for the analysis of an OTU table generated from 16S sequencing. It also includes the data from the North Temperate Lakes-Microbial Observatory used in the paper "".
Three data files are included: otu_table, taxonomy, and metadata. Access these by calling them with data().
}
\details{
\tabular{ll}{
Package: \tab OTUtable\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-29\cr
License: \tab GPL-2 \cr
}
Functions include:
bog_subset
chao1
clean_shared
clean_taxonomy
combine_otus
expand_taxa
extract_date
grab_group
make_do_matrix
make_temp_matrix
obs_richness
pielou
plot_column
reduce_names
remove_reps
rotate
shannon
strat_metric
year_subset
zscore

}
\author{
Alexandra Linz <amlinz16@gmail.com>
}
\references{
Manuscript in preparation
}
\keyword{ package }

