\name{OTUtable-package}
\alias{OTUtable-package}
\alias{OTUtable}
\docType{package}
\title{
OTU table analysis functions
}
\description{
This package contains functions for the analysis of an OTU table generated from 16S sequencing. It also includes the data from the North Temperate Lakes-Microbial Observatory used in the paper "".
Three data files are included: otu_table, taxonomy, and metadata. Access these by calling them with data().
}
\details{
\tabular{ll}{
Package: \tab OTUtable\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2016-03-11\cr
License: \tab GPL-3 \cr
}
Functions include:
bog_subset \cr
chao1 \cr
clean_shared \cr
clean_taxonomy \cr
combine_otus \cr
extract_date \cr
grab_group \cr
make_do_matrix \cr
make_temp_matrix \cr
obs_richness \cr
pielou \cr
plot_column \cr
reduce_names \cr
remove_reps \cr
rotate \cr
shannon \cr
strat_metric \cr
year_subset \cr
zscore \cr
}
\author{
Alexandra Linz <amlinz16@gmail.com>
}
\references{
Linz, A.M., Crary, B., Shade, A., Herren, C.M., Owens, S., Gilbert, J.A., Knight, R., and McMahon, K.D. (2016) "Seasonal mixing as a barrier to the development of stable bacterial communities in bog lakes." Manuscript in prep. 
}
\keyword{ package }

